# Parse the standard library at most twice for host == target