--- a/Sources/Driver/Driver.swift
+++ b/Sources/Driver/Driver.swift
@@ -157,7 +157,7 @@ public struct Driver: ParsableCommand {
     let productName = makeProductName(inputs)
 
     /// An instance that includes just the standard library.
-    var ast = freestanding ? Host.freestandingLibraryAST : Host.standardLibraryAST
+    var ast = freestanding ? Host.freestandingLibraryAST : Host.hostedLibraryAST
 
     // The module whose Hylo files were given on the command-line
     let sourceModule = try ast.makeModule(