--- a/Tests/HyloTests/LoweringTests.swift
+++ b/Tests/HyloTests/LoweringTests.swift
@@ -14,7 +14,8 @@ extension XCTestCase {
     try checkAnnotatedHyloFileDiagnostics(inFileAt: hyloFilePath, expectSuccess: expectSuccess) {
       (valSource, diagnostics) in
       // Note: built-in module is visible so that we can test built-in function calls.
-      var ast = AST(libraryRoot: coreLibrarySourceRoot, for: CompilerConfiguration())
+      var ast = Host.freestandingLibraryAST
+
       let module = try ast.makeModule(
         valSource.baseName, sourceCode: [valSource], builtinModuleAccess: true,
         diagnostics: &diagnostics)