--- a/Tests/ManglingTests/ManglingTests.swift
+++ b/Tests/ManglingTests/ManglingTests.swift
@@ -3,6 +3,7 @@ import FrontEnd
 import IR
 import StandardLibrary
 import TestUtils
+import Utils
 import XCTest
 
 final class ManglingTests: XCTestCase {
@@ -67,7 +68,7 @@ final class ManglingTests: XCTestCase {
 
     let input = SourceFile(synthesizedText: text, named: "main")
     let (p, m) = try checkNoDiagnostic { (d) in
-      var ast = AST(libraryRoot: standardLibrarySourceRoot, for: CompilerConfiguration())
+      var ast = Utils.Host.standardLibraryAST
       let main = try ast.makeModule("Main", sourceCode: [input], diagnostics: &d)
       let base = ScopedProgram(ast)
       return (try TypedProgram(annotating: base, reportingDiagnosticsTo: &d), main)
@@ -94,7 +95,7 @@ final class ManglingTests: XCTestCase {
 
   func testTypes() throws {
     let p = try checkNoDiagnostic { (d) in
-      let ast = AST(libraryRoot: standardLibrarySourceRoot, for: CompilerConfiguration())
+      let ast = Host.standardLibraryAST
       let base = ScopedProgram(ast)
       return try TypedProgram(annotating: base, reportingDiagnosticsTo: &d)
     }