In an attempt to speed up testing, instead of parsing the library each time we need an AST for it as a starting point for testing, store hosted and/or freestanding ASTs in global constants, which are initialized once, lazily and thread-safely in Swift.

Also fixes #932.

## Results

- Debug:
   - Parallel testing (`time swift test --parallel --skip-build`):
      - Before: `440.54s user 23.45s system 834% cpu 55.578 total`
      - After:    `441.36s user 23.85s system 834% cpu 55.723 total`
   - Serial testing  (`time swift test --parallel --skip-build`):
      - Before: `325.20s user 11.28s system 96% cpu 5:49.06 total`
      - After:    `289.28s user 10.43s system 95% cpu 5:13.35 total`
- Release:
   - Parallel testing (`time swift test --parallel --skip-build`):
      - Before: `263.19s user 29.15s system 618% cpu 47.302 total`
      - After:    `264.44s user 28.92s system 624% cpu 46.998 total`
   - Serial testing  (`time swift test --parallel --skip-build`):
      - Before: `211.38s user 15.21s system 144% cpu 2:36.36 total`
      - After:   `186.19s user 14.69s system 155% cpu 2:09.02 total`

The `time` man page is unhelpful but from my experiment (`time sleep 5` yields `0.00s user 0.00s system 0% cpu 5.014 total`), the number we mostly care about is the last one, total.

This change seems to make zero difference if parallel testing is enabled, for reasons I don't understand, and the wins from parallel testing for a developer far outweigh any improvement we get from this change when testing serially.  I think it's still a change worth making.