@kyouko-taiga 2023-12-08
I haven't carefully reviewed the PR yet but I agree that it seems like a change worth making. As I mentioned on Slack, we can probably perform type checking only once (or twice as the title suggest?) for all the tests, which is significantly more expensive than parsing. So this PR is a very nice step toward the hopefully bigger improvement we'll get from reusing the result of type checking.

For information, parsing is about 15% of the total exec time when compiling with `--typecheck` in release, which seems to explain the numbers reported in the PR description.