@kyouko-taiga 2023-12-08
Maybe `--parallel` forks the test driver before it got a change to initialize your static lets?