@dabrahams 2023-12-08
I never considered that parallel was done by forking rather than by launching threads, but now that you mention it, I'm sure that's the case, since it's very likely the average Swift programmer is testing thread-unsafe constructs (to say nothing of XCTest itself).  Still, to have zero impact it would have to be forking newly for each test.  If that's the case, maybe I can find some kind of testing setup hook that will let me do the initialization earlier.