@dabrahams 2023-12-08
Nope, it's [using threads](https://github.com/apple/swift-package-manager/blob/main/Sources/Commands/SwiftTestTool.swift#L929).  And not even using Dispatch, which is a bit shocking!